test: claim canonical JSON in production trace contract
